Make shared rerun button
