+++ app/src/ui/check-runs/ci-check-run-popover.tsx
@@ -11,7 +11,6 @@ import {
   FailingCheckConclusions,
 } from '../../lib/ci-checks/ci-checks'
 import { Octicon, syncClockwise } from '../octicons'
-import { Button } from '../lib/button'
 import { APICheckConclusion, IAPIWorkflowJobStep } from '../../lib/api'
 import { Popover, PopoverCaretPosition } from '../lib/popover'
 import { CICheckRunList } from './ci-check-run-list'
@@ -21,7 +20,7 @@ import * as OcticonSymbol from '../octicons/octicons.generated'
 import { Donut } from '../donut'
 import { supportsRerunningChecks } from '../../lib/endpoint-capabilities'
 import { getPullRequestCommitRef } from '../../models/pull-request'
-import { IMenuItem, showContextualMenu } from '../../lib/menu-item'
+import { CICheckReRunButton } from './ci-check-re-run-button'
 
 const BlankSlateImage = encodePathAsUrl(
   __dirname,
@@ -233,29 +232,11 @@ export class CICheckRunPopover extends React.PureComponent<
     }
 
     return (
-      <Button
-        onClick={this.onRerunChecksButton}
+      <CICheckReRunButton
         disabled={checkRuns.length === 0 || this.state.loadingActionWorkflows}
-      >
-        <Octicon symbol={syncClockwise} /> Re-run checks{' '}
-        <Octicon symbol={OcticonSymbol.triangleDown} />
-      </Button>
+        onRerunChecks={this.rerunChecks}
+      />
     )
-  }
-
-  private onRerunChecksButton = () => {
-    const items: IMenuItem[] = [
-      {
-        label: __DARWIN__ ? 'Re-run Failed Checks' : 'Re-run failed checks',
-        action: () => this.rerunChecks(true),
-      },
-      {
-        label: __DARWIN__ ? 'Re-run All Checks' : 'Re-run all checks',
-        action: () => this.rerunChecks(false),
-      },
-    ]
-
-    showContextualMenu(items)
   }
 
   private renderCheckRunLoadings(): JSX.Element {
